7. conflict + final merging

+++ src/SalesManager.java
@@ -15,18 +15,18 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = sales[0];
-        for (int sale : sales) {
+    public long min() {
+        long min = sales[0];
+        for (long sale : sales) {
             if (sale < min)
                 min = sale;
         }
         return min;
     }
 
-    public int average() {
-        int sum = 0;
-        for (int sale : sales) {
+    public long average() {
+        long sum = 0;
+        for (long sale : sales) {
             sum += sale;
         }
         return (sum - max() - min()) / (sales.length - 2);
